rename one more class
